+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - Deprecate Aqua (QA) as a target IntelliJ Platform
 - Print requested IntelliJ Platform when throwing exception about unresolved dependency.
+- Make `IntelliJPlatformDependenciesHelper` aware of custom IntelliJ Platform when used within custom tasks.
 
 ### Fixed